type cleanups. reduces memory usage on 64bit, no difference on 32bit. closes #780
svn path=/icecast/trunk/icecast/; revision=14306

+++ src/client.h
@@ -42,7 +42,7 @@ typedef struct _client_tag
     refbuf_t *refbuf;
 
     /* position in first buffer */
-    unsigned long pos;
+    unsigned int pos;
 
     /* auth used for this client */
     struct auth_tag *auth;
